requirements.txt
django>1.9,<1.10
djangorestframework
markdown
django-filter